revert(snapshot): keep Snapshot model class client-side only
Hoie 2026-05-15: not using model classes on the server is a deliberate
design choice; reverting the move into common. Backfill goes back to a
plain JSON-typed snapshot literal.
