+++ src/server/lib/compute-tools/backfill-snapshots.ts
@@ -3,7 +3,6 @@ import {
   getSquashedDateString,
   getYearMonthString,
   JSONSecuritySnapshot,
-  Snapshot,
 } from "common";
 import {
   getSecuritySnapshots as realGetSecuritySnapshots,
@@ -164,10 +163,10 @@ export const backfillMonthlySecuritySnapshotsForward = async (
       }
 
       const newSnapshot: JSONSecuritySnapshot = {
-        snapshot: new Snapshot({
+        snapshot: {
           snapshot_id: `${security_id}-${getSquashedDateString(new Date(dayInMonth))}`,
           date: new Date(dayInMonth).toISOString(),
-        }),
+        },
         security: {
           security_id,
           ticker_symbol,
